Refactor, reorganize, reformat, and some cleanup since we're getting close to "full" release
