--- examples/display.rs
@@ -1,1 +1,0 @@
-// this will eventually contain an example of what is to be done with the library that is created
