+++ examples/old_tests/get_camera_properties.rs
