+++ examples/old_tests/snap_single_image.rs
