--- src/components.rs
@@ -1,1 +1,0 @@
-// this shall be left blank for now, but later removed most likely
