+++ src/lib.rs
@@ -1,13 +1,15 @@
+pub mod audio;
 pub mod bluetooth;
 pub mod camera;
+pub mod gui;
 pub mod plugin;
 pub use nokhwa;
-mod systems;
+mod zoetrope;
 
 pub mod prelude {
     pub use crate::{
         camera::{VideoFrame, VideoStream},
+        gui::{open_window, ui_test, UiState},
         plugin::ZoetropePlugin,
-        systems::{UiState,ui_test, open_window},
     }; // temporary names for right now, these may change based on the current needs of the project
 }
