+++ src/plugin.rs
@@ -3,7 +3,7 @@ use bevy_tokio_tasks::TokioTasksPlugin;
 
 use crate::bluetooth::{async_spawner, ZoetropeRotation};
 use crate::camera::VideoFrame;
-use crate::systems::{logical_camera_rotation, physical_camera_setup, update_zoetrope_image};
+use crate::zoetrope::{logical_camera_rotation, update_zoetrope_image, zoetrope_setup};
 
 pub struct ZoetropePlugin;
 
@@ -12,7 +12,7 @@ impl Plugin for ZoetropePlugin {
         app.insert_resource(VideoFrame(Handle::default()))
             .add_plugin(TokioTasksPlugin::default())
             .insert_resource(ZoetropeRotation(0))
-            .add_startup_system(physical_camera_setup)
+            .add_startup_system(zoetrope_setup)
             .add_startup_system(async_spawner)
             .add_system(bevy::window::close_on_esc)
             .add_system(update_zoetrope_image)
